pyproject.toml
[project]
name = "run-pipeline"
dynamic = ["version"]
requires-python = ">=3.11"
description = "Pipeline runner with uv groups for per-segment envs"
readme = "README.md"
license = "MIT"

# Example add new package: uv add --group drift-correct "scikit-image>=0.19.0" 

[tool.setuptools]
packages = ["standard_code", "standard_code.python", "standard_code.python.drift_correction_utils", "standard_code.python.bioio_imaris"]

[tool.setuptools.dynamic]
version = {file = "VERSION"}

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
# Intentionally empty; uv groups are defined below under [tool.uv.dependency-groups]

[tool.uv]
# Consider best match across indexes when multiple are configured
index-strategy = "unsafe-best-match"

# Use the PyTorch CUDA index only for torch packages

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }
torchaudio = { index = "pytorch" }

 [dependency-groups]
convert-to-tif = [
	"numpy",
	"matplotlib",
	"tqdm",
	"joblib",
	"nd2reader",
	"scyjava",
	"jpype1==1.5.0",
	"dask[array]",
	"h5py",
	"imaris-ims-file-reader",
	"pystackreg",
	"cupy-cuda12x",
	"bioio",
	"bioio-bioformats",
	"bioio-ome-tiff",
	"bioio-tifffile",
	"bioio-nd2",
	"bioio_lif",
	"bioio_czi",
	"bioio_dv",
	"pyGPUreg",
]
drift-correct = [
	"numpy",
	"matplotlib",
	"tqdm",
	"joblib",
	"nd2reader",
	"scyjava",
	"jpype1==1.5.0",
	"dask[array]",
	"h5py",
	"imaris-ims-file-reader",
	"pystackreg",
	"cupy-cuda12x",
	"bioio",
	"bioio-bioformats",
	"bioio-ome-tiff",
	"bioio-tifffile",
	"bioio-nd2",
	"bioio_lif",
	"bioio_czi",
	"bioio_dv",
	"pyGPUreg",
	"imreg_dft",
	"scikit-image>=0.19.0",
	"scikit-learn>=1.1",
	"pandas",
	"scipy",
	"opencv-python>=4.11.0.86",
]
segment-ernet = [
	"torch==2.4.0+cu124",
	"torchvision==0.19.0+cu124",
	"torchaudio==2.4.0+cu124",
	"python-dotenv",
	"tqdm",
	"numpy>=1.24,<2.0",
	"numba",
	"scipy",
	"scikit-image",
	"opencv-python",
	"matplotlib",
	"pillow",
	"networkx<3",
	"imagecodecs",
	"ipykernel",
	"timm",
	"einops",
	"sknw==0.12",
	"bioio",
	"bioio-bioformats",
	"bioio-ome-tiff",
	"bioio-tifffile",
	"bioio-nd2",
	"bioio_lif",
	"bioio_czi",
	"bioio_dv",
]
mask_measure = [
	"numpy",
	"pandas",
	"scikit-image",
	"tqdm",
	"bioio",
	"bioio-bioformats",
	"bioio-ome-tiff",
	"bioio-tifffile",
	"bioio-nd2",
	"bioio_lif",
	"bioio_czi",
	"bioio_dv",
]
plot = [
	"numpy",
	"matplotlib",
	"pandas",
	"seaborn",
	"bioio",
	"bioio-bioformats",
	"bioio-ome-tiff",
	"bioio-tifffile",
	"bioio-nd2",
	"bioio_lif",
	"bioio_czi",
	"bioio_dv",
	"scipy",
	"scikit-image",
	"scipy",
]
segmentation = [
	"numpy",
	"scipy",
	"scikit-image",
	"pandas",
	"trackpy",
	"opencv-python>=4.11.0.86",
	"tqdm",
	"bioio",
	"bioio-bioformats",
	"bioio-ome-tiff",
	"bioio-tifffile",
	"bioio-nd2",
	"bioio_lif",
	"bioio_czi",
	"bioio_dv",
	"joblib",
	"nd2reader",
	"seaborn",
	"roifile",
	"h5py",
]
denoise = [
	"numpy",
	"opencv-python>=4.11.0.86",
	"scikit-image>=0.19.0",
	"n2v>=0.3.2",
	"tensorflow>=2.10.0,<2.16.0",
	"keras<3.0",
	"bioio",
	"bioio-bioformats",
	"bioio-ome-tiff",
	"bioio-tifffile",
	"bioio-nd2",
	"bioio_lif",
	"bioio_czi",
	"bioio_dv",
	"tqdm",
]
run_ilastik = [
	"numpy",
	"scipy",
	"scikit-image",
	"matplotlib",
	"opencv-python",
	"roifile",
	"trackpy",
	"filelock",
	"h5py",
	"tqdm",
	"joblib",
	"pyyaml",
	"bioio",
	"bioio-bioformats",
	"bioio-ome-tiff",
	"bioio-tifffile",
	"bioio-nd2",
	"bioio_lif",
	"bioio_czi",
	"bioio_dv",
	"nd2reader",

]

imagej = [
	"pyimagej>=1.4.0",
	"scyjava",
	"jpype1==1.5.0",
	"numpy",
	"bioio",
	"bioio-bioformats",
	"bioio-ome-tiff",
	"bioio-tifffile",
	"bioio-nd2",
	"bioio_lif",
	"bioio_czi",
	"bioio_dv",
	"tqdm",
	"matplotlib"
]

# Development dependencies
dev = [
	"pytest>=7.0",
	"pytest-cov",
	"black",
	"isort",
	"flake8",
	"mypy",
	"pre-commit",
	"jupyter",
	"ipython",
]
